moveoutlist/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
